w3c-reports/12092014/mac-wk2/runned-tests/XMLHttpRequest/send-network-error-async-events.sub-expected.txt

CONSOLE MESSAGE: XMLHttpRequest cannot load http://nonexistent-origin.localhost:8800/. Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin.

PASS XmlHttpRequest: The send() method: Fire a progress event named error when Network error happens (synchronous flag is unset) 
